requirements.txt
Django==2.2.2
psycopg2==2.8.3
django-dynamic-fixture==2.0.0
django-test-plus==1.0.22